release: 130002.9.0 (based on 13.0.2, version 9.0)
add a 'type' attribute to 'middleware' rules
to differenciate the grouping of 'content-type' values
to which the rule should apply.

previously, all 'middleware' assumed that the response was html.

now, the following 'type' values are supported:
  'html', 'json', 'js', 'text'

while 'js' and 'text' middleware are both passed
the same unmodified response text,
distinguishing these as separate 'type' values
allows a 'middleware' rule to target one but not the other.