release: 130002.11.0 (based on 13.0.2, version 11.0)
the default status code for redirect responses
is conditional on the request method:
* 307: POST, PUT, PATCH
* 301: otherwise

a 301 redirect response always causes the client
to follow the redirect with a GET request;
any data included in the body of the original request is lost.

a 307 redirect response always causes the client
to follow the redirect with the original request method;
any data included in the body of the original request is preserved.

previously, the default status code was always 301.

this update only changes the default status code,
but does not effect the ability for redirect rules
to override the default status code on a per-rule basis.