release: 130002.13.1 (based on 13.0.2, version 13.1)
adds support for cgi-bin scripts to process POST requests.
POST data is passed to the subprocess through its standard input.

adds more granular control over when to return: 405 Method Not Allowed

rewrites the "slasher" library;
mainly for the purpose to normalize the result to a posix path.
presently, this libarary is only used to prepare paths for minimatch.